double click to set is now working
