+++ src/components/nodes/DataNode.tsx
@@ -4,14 +4,19 @@ import { CustomNodeProps } from "@src/types/node";
 import NodeWrapper from "@/components/common/NodeWrapper";
 import HandleComponent from "@/components/common/HandleComponent";
 import { textWrap } from "@src/utils/TextWrap";
+import { useFlowChartGraph } from "@hooks/useFlowChartGraph";
 
 const DataNode = (props: CustomNodeProps) => {
   const {
     nodeProps: { data },
     isRunning,
     nodeError,
+    node,
   } = props;
   const [isRenamingTitle, setIsRenamingTitle] = useState(false);
+  const [newTitle, setNewTitle] = useState(data.label);
+  const { handleTitleChange } = useFlowChartGraph();
+
   return (
     <NodeWrapper wrapperProps={props}>
       <div
@@ -27,11 +32,15 @@ const DataNode = (props: CustomNodeProps) => {
         {isRenamingTitle ? (
           <input
             type="text"
-            value={data.label}
+            value={newTitle}
             onKeyDown={(event) => {
-              if (event.key === "enter") {
+              if (event.key === "Enter") {
                 setIsRenamingTitle(false);
               }
+            }}
+            onChange={(e) => {
+              setNewTitle(e.target.value);
+              handleTitleChange(e.target.value, node);
             }}
           />
         ) : (
