+++ src/feature/flow_chart_panel/components/node-edit-menu/NodeEditMenu.tsx
@@ -6,8 +6,6 @@ import NodeEditModal from "./NodeEditModal";
 type NodeEditMenuProps = {
   selectedNode: Node<ElementsData> | null;
   unSelectedNodes: Node<ElementsData>[] | null; //used in ParamField.tsx for references
-  nodes: Node<ElementsData>[];
-  setNodes: (nodes: Node<ElementsData>[]) => void;
   setNodeModalOpen: (open: boolean) => void;
   handleDelete: (nodeId: string, nodeLabel: string) => void;
 };
@@ -15,8 +13,6 @@ type NodeEditMenuProps = {
 export const NodeEditMenu = ({
   selectedNode,
   unSelectedNodes,
-  nodes,
-  setNodes,
   setNodeModalOpen,
   handleDelete,
 }: NodeEditMenuProps) => {
@@ -28,8 +24,6 @@ export const NodeEditMenu = ({
         <NodeEditModal
           node={selectedNode}
           otherNodes={unSelectedNodes}
-          nodes={nodes}
-          setNodes={setNodes}
           setNodeModalOpen={setNodeModalOpen}
           handleDelete={handleDelete}
         />
